2020-10-20 오후 5:51 각 패키지 안에 주석 추가, dao의 누락된 코딩 추가, 전체 방문자 구현을 위한 StatsDao, StatsService 등에 코드 추가.

+++ sakila/src/sakila/vo/Stats.java
@@ -2,17 +2,17 @@ package sakila.vo;
 
 public class Stats {
 	private String day;
-	private long count;
+	private int count;
 	public String getDay() {
 		return day;
 	}
 	public void setDay(String day) {
 		this.day = day;
 	}
-	public long getCount() {
+	public int getCount() {
 		return count;
 	}
-	public void setCount(long count) {
+	public void setCount(int count) {
 		this.count = count;
 	}
 	@Override
